+++ pya/aserver.py
@@ -4,6 +4,7 @@ import logging
 import numpy as np
 from warnings import warn
 
+
 _LOGGER = logging.getLogger(__name__)
 _LOGGER.addHandler(logging.NullHandler())